ra_update_linhas_kml/requirements.txt
pandas
numpy
python-dotenv
click
sqlalchemy
requests
psycopg2-binary
geopandas
shapely
geojson
protobuf
xlsxwriter
geojson-length
fastkml
lxml
